Core Changes
    New Feature: Text Analysis Mode: Implemented a new text mode capable of analyzing entire texts. This lays the groundwork for future advanced linguistic processing, potentially integrating tools like Zemberek for operating on tokenized texts.

    Module Restructuring: Significantly refactored and decomposed interactive_trainer.py. The original file was broken down into three distinct, more manageable modules.

    Class Refinements: The decompose function was made a method within the InteractiveTrainer class, improving encapsulation and object-oriented design.

Optimization & Future Work

    Decomposition Rule Optimization: Added new suffix rules that substantially reduced the number of "possible decompositions," leading to improved performance and precision. More rules will be added in subsequent commits.

    Suffix Class Enhancement (Next Step): Introduced several new suffixes. However, due to the polymorphic nature of these suffixes (e.g., they behave differently in various contexts), the Suffix class will require additional information or a structural update in the near future to handle this complexity effectively.
