+++ data/config.py
@@ -4,6 +4,7 @@ class TrainingConfig:
     """Configuration for the interactive trainer"""
     model_path: str = "data/turkish_morph_model.pt"
     vocab_path: str = "data/suffix_vocab.json"
+    words_path: str = "data/words.txt"
     training_count_file: str = "data/training_count.txt"
     valid_decompositions_file: str = "data/valid_decompositions.jsonl"
     checkpoint_frequency: int = 10  # Save every N examples
