+++ data/sample.txt
@@ -1,0 +1,1 @@
+Bu nedenle bilimle uğraşanların buraya gelmek istemeleri doğaldı. Ayrıca Bağdat bu dönemde Türk nüfuzu altına girmiş ve halifeleri ayakta tutan da buradaki Türklerdi
