+++ data/training_count.txt
@@ -1,1 +1,1 @@
-544
+1247
