+++ util/word_methods.py
@@ -5,11 +5,21 @@ from pathlib import Path
 DATA_FILE = Path(__file__).resolve().parent.parent / "data" / "words.txt"
 
 
+## Vowel Classes
+BACK_FLAT   = ['a','ı']
+BACK_ROUND  = ['o','u']
 
-# --- Constants ---
-VOWELS = ['a','o','u','ı','e','ö','ü','i']
-BACK_VOWELS = ['a','ı','o','u']
-FRONT_VOWELS = ['e','i','ö','ü']
+FRONT_FLAT  = ['e','i']
+FRONT_ROUND = ['ö','ü']
+
+BACK_VOWELS  = BACK_FLAT  + BACK_ROUND
+FRONT_VOWELS = FRONT_FLAT + FRONT_ROUND 
+
+VOWELS = BACK_VOWELS + FRONT_VOWELS
+
+
+
+
 HARD_CONSONANTS = ['f','s','t','k','ç','ş','h','p']  # fıstıkçı şahap
 
 
@@ -54,6 +64,42 @@ def exists(word: str) -> bool:
 
     return False
 
+## delete from words.txt
+def delete(word: str) -> bool:
+    """
+    Deletes the given word from the file 'words.txt'.
+    Returns True if the word was found and deleted, False otherwise.
+    """
+    try:
+        # Read all words (strip whitespace)
+        
+        # Check if the word exists
+        if word not in WORDS:
+            return False
+        
+        # Remove the word and rewrite the file
+        WORDS.remove(word)
+        with open(DATA_FILE, "w",encoding="utf-8") as file:
+            for w in WORDS:
+                file.write(w + "\n")
+        
+        return True
+
+    except FileNotFoundError:
+        print("Error: words.txt not found.")
+        return False
+    except Exception as e:
+        print(f"An error occurred: {e}")
+        return False
+    
+def random_word() -> str:
+    return random.choice(list(WORDS))
+
+
+def can_be_verb(word: str) -> bool:
+    """Checks if a root is a verb by verifying its infinitive form."""
+    return exists(infinitive(word))
+
 
 # --- Harmony functions ---
 def major_harmony(word: str) -> MajorHarmony | None:
@@ -87,10 +133,6 @@ def infinitive(word: str) -> str:
     return word + suffix
 
 
-def can_be_verb(word: str) -> bool:
-    """Checks if a root is a verb by verifying its infinitive form."""
-    return exists(infinitive(word))
-
 
 def ends_with_vowel(word: str) -> bool:
     """Check if word ends with a vowel"""
@@ -111,32 +153,3 @@ def has_no_vowels(word: str) -> bool:
             break
     return ret
 
-def delete(word: str) -> bool:
-    """
-    Deletes the given word from the file 'words.txt'.
-    Returns True if the word was found and deleted, False otherwise.
-    """
-    try:
-        # Read all words (strip whitespace)
-        
-        # Check if the word exists
-        if word not in WORDS:
-            return False
-        
-        # Remove the word and rewrite the file
-        WORDS.remove(word)
-        with open(DATA_FILE, "w",encoding="utf-8") as file:
-            for w in WORDS:
-                file.write(w + "\n")
-        
-        return True
-
-    except FileNotFoundError:
-        print("Error: words.txt not found.")
-        return False
-    except Exception as e:
-        print(f"An error occurred: {e}")
-        return False
-    
-def random_word() -> str:
-    return random.choice(list(WORDS))
